Master File Submitted
From here on out I'd suggest not modifying the "Master Folder" since it
has everything we have up to this point. So we have a backup. Until we
figure out that other editing thing just edit the normal "RailAvoid"
folder when you work on the game and leave "RailAvoideMaster" Untouched.