Health Bar Fix and New Folder Update
I made a copy of RailAvoid since I was working on it and made the copy
under my name. It'll be the most updated on some of the coding I do. At
least this way there are some copies and stuff that hopefully won't
vanish. I also have my copy on my desktop just in case. But this is also
here for you guys to see what's been done on my side.