DanM Update
I nearly broke the damn thing on my side when I tried to sync things up
so I'm taking a break for now.

The good news is a lot's done:
++Enemy's movement AI is good to go, relatively. They fly up to player,
stop for a malleable amount of time, then fly past the player. No
Shooting Yet.
++Enemy Shot Types are built and can be edited. They just don't fire
normally and are just in the field to be seen.
++All known Rings work, White is additional Score, Yellow speeds up,
Blue Speeds Down, and Green Heals.
++HealthBar works perfectly for now. If we want to change its colors or
something then that's another story.
++New Particle Effects are builts, such as a better looking jet fire
animation and a smoother explosion animation... however making scripts
to have any of these emit on command is another story atm.
++Basic Asteroid Patterns are built to be randomly generated among the
rest of the stuff.

Things missing still
--Enemies shooting
--Level progression
--Score multiplier conditions
--Comets need Fixing
--Certain Enemy Fire Behavior needs fixing (such as the big lasers
Yellow and indigo)
--Making Particles emit on command (like a small explosion when hurt)
--Options, and similar UI elements

And I'm done for now. I didn't commit the changes so maybe some of the
missing bits are filled in already. I dunno. I'll check later.