Crisis Averted, More Updates
I synced up my folder with the latest updates, so the main menu and
enemy ships are in and implemented. However the UI for the main menu
doesn't seem to be working (I did set up a build. I even tested it on
the other version and no buttons came up) so we might need to go over
this live on Wednesday.

Other than that, the issues I had with the earlier sync up seems to be
resolved and nothing seems to be missing. I'm sorry for the scare.